--- server/agent/src/main/java/ai/starwhale/mlops/agent/task/action/Condition.java
@@ -1,13 +1,0 @@
-/*
- * Copyright 2022.1-2022
- * StarWhale.ai All right reserved. This software is the confidential and proprietary information of
- * StarWhale.ai ("Confidential Information"). You shall not disclose such Confidential Information and shall use it only
- * in accordance with the terms of the license agreement you entered into with StarWhale.com.
- */
-
-package ai.starwhale.mlops.agent.task.action;
-
-public interface Condition<Input> {
-
-    boolean judge(Input input);
-}
