--- server/agent/src/main/java/ai/starwhale/mlops/agent/task/action/SelectOneToExecute.java
@@ -1,20 +1,0 @@
-/*
- * Copyright 2022.1-2022
- * StarWhale.ai All right reserved. This software is the confidential and proprietary information of
- * StarWhale.ai ("Confidential Information"). You shall not disclose such Confidential Information and shall use it only
- * in accordance with the terms of the license agreement you entered into with StarWhale.com.
- */
-
-package ai.starwhale.mlops.agent.task.action;
-
-public interface SelectOneToExecute<Input, Output> {
-
-    default void apply(Input input, Context context, Condition<Input> condition,
-        DoTransition<Input, Output> one, DoTransition<Input, Output> another) {
-        if (condition.judge(input)) {
-            one.apply(input, context);
-        } else {
-            another.apply(input, context);
-        }
-    }
-}
